Ajout de la config sécurité JWT/CORS, OpenAPI/Swagger et properties prod. Nouvelles entités/migrations (audit conges, jours fériés) + tests d’intégration/uni. Docs complètes (quick start, sécurité, tests, améliorations, checklists). Frontend congés : dashboard, hook d’erreurs API, modèles de types/solde mis à jour.

+++ GestionRH/src/main/java/com/fares/gestionrh/dto/conge/CongeResponse.java
@@ -21,7 +21,7 @@ public class CongeResponse {
     private String statut;
     private String motif;
     private String commentaireValidation;
-    private Long nombreJours;
+    private Double nombreJours;
 
     // Informations de l'employé
     private Long employeId;
